requirements.txt
streamlit==1.44.1
python-docx==1.1.2
PyPDF2==3.0.1
langchain==0.3.23
langchain-community==0.3.21
huggingface_hub==0.30.2
openai==1.72.0
toml==0.10.2
